+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Added**
 
+* Interface for fetching samples given a project code and sample status
+
 **Fixed**
 
 **Dependencies**