+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Added**
 
+* Use case counting all samples of a project and the subset of samples having been received at the lab
+
 * Interface for counting samples and received samples given a project code
 
 * Add resource service for project samples status list. (`#34 <https://github.com/qbicsoftware/sample-tracking-status-overview/pull/34>`_)